fix: Code splitting
